[refactor] split RangeIndexConfigCondition into abstract base class and an attribute condition implementation
and move predicate analyzation from pragma to the condition itself
